# Reimplement datediff and dateadd in c to improve performance