--- a/contrib/babelfishpg_tsql/src/err_handler.c
+++ b/contrib/babelfishpg_tsql/src/err_handler.c
@@ -131,7 +131,10 @@ is_ignorable_error(int pg_error_code, uint8_t override_flag)
 		case SQL_ERROR_8145:
 		case SQL_ERROR_8146:
 		case SQL_ERROR_213:
-		case SQL_ERROR_6615:			
+		case SQL_ERROR_6615:
+		case SQL_ERROR_155:
+		case SQL_ERROR_9810:
+		case SQL_ERROR_535:
 		case SQL_ERROR_15003:
 			{
 				elog(DEBUG1, "TSQL TXN is_ignorable_error %d", latest_error_code);