--- a/test/JDBC/expected/babel_datetime-vu-verify.out
+++ b/test/JDBC/expected/babel_datetime-vu-verify.out
@@ -376,8 +376,7 @@ select dateadd(year, 150, cast('9900-12-26 23:29:29' as datetime))
 go
 ~~ERROR (Code: 517)~~
 
-~~ERROR (Message: data out of range for datetime)~~
-
+~~ERROR (Message: Adding a value to a 'datetime' column caused an overflow.)~~
 
 -- Test data type precedence TODO Fix [BABEL-883] missing TDS support for type regtype (was pg_typeof produces error in sqlcmd)
 select pg_typeof(c1) FROM (SELECT cast('2016-12-26 23:30:05' as datetime) as C1 UNION SELECT cast('2016-12-26 23:30:05' as smalldatetime) as C1) T