--- a/test/JDBC/upgrade/14_10/schedule
+++ b/test/JDBC/upgrade/14_10/schedule
@@ -391,7 +391,6 @@ BABEL-3914
 BABEL_OBJECT_NAME
 sys-systypes
 BABEL_OBJECT_DEFINITION
-datediff_internal_date
 Test-sp_rename
 Test-sp_rename-dep
 BABEL-3657